requirements.txt
easyocr==1.7.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
tqdm==4.66.5
PyYAML==6.0.2
rapidfuzz==3.9.7
